refactor function manager: remove keep order, add retry, add test (#127)

+++ protocol/mqtt/config.go
@@ -8,7 +8,7 @@ import (
 
 // TopicInfo with topic and qos
 type TopicInfo struct {
-	QOS   byte   `yaml:"qos" json:"qos" default:"0" validate:"min=0, max=1"`
+	QOS   uint32 `yaml:"qos" json:"qos" default:"0" validate:"min=0, max=1"`
 	Topic string `yaml:"topic" json:"topic" validate:"nonzero"`
 }
